solucion problemas al cargar las listas y borrar o editar datos

+++ movie-service/src/main/java/com/movie/service/repositorio/DirectorRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.movie.service.Entidades.Director;
 
-public interface DirectorRepository extends JpaRepository<Director, Integer> {
+public interface DirectorRepository extends JpaRepository<Director, Long> {
 	
 }
